Add mockup link

+++ priv/repo/migrations/20210704081108_create_link_details.exs
@@ -3,7 +3,7 @@ defmodule Linkir.Repo.Migrations.CreateLinkDetails do
 
   def change do
     create table(:link_details) do
-      add :price, :float
+      add :price, :integer
       add :raw_response, :text
       add :crawled_at, :naive_datetime
       add :link_id, references(:links, on_delete: :nothing)
